requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
black==21.5b2
click==8.0.1
decorator==5.0.9
Django==3.2.3
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
